+++ execution_engine/src/core/resolvers/v1_function_index.rs
@@ -60,6 +60,7 @@ pub(crate) enum FunctionIndex {
     RandomBytes,
     DictionaryReadFuncIndex,
     EnableContractVersion,
+    GenericHash,
 }
 
 impl From<FunctionIndex> for usize {